Feature 3/handling crud operations

+++ app/Http/Controllers/BrandController.php
@@ -18,17 +18,42 @@ class BrandController extends Controller
      */
     public function __construct() {
         // Functions inside the authentication controller can not be accessed without having the valid token.
-        $this->middleware('auth:api');
+        $this->middleware('jwt');
     }
     
     /**
-     * Display a listing of the resource.
-     * 
-     * @author Nderi Kamau <nderikamau1212@gmail.com>
+     * @OA\Get(
+     *      path="/api/v1/brands",
+     *      operationId="getBrandsList",
+     *      summary="Get list of Brands",
+     *      description="Returns list of brands",
+     *      tags={"Brands"},
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *      	  @OA\JsonContent(
+     *              type="object",
+     *              @OA\Property(
+     *                  property="brands",
+     *                  collectionFormat="multi"
+     *              )
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=404,
+     *          description="Brands not found"
+     *      )
+     * )
      */
     public function index()
     {
         $brands = Brand::all();
+
+        if (!$brands) {
+            return response()->json([
+                'error' => 'Brands not found'
+            ], Response::HTTP_NOT_FOUND);
+        }
 
         // Return response with message and data
         return response()->json([
@@ -37,9 +62,41 @@ class BrandController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
-     * 
-     * @author Nderi Kamau <nderikamau1212@gmail.com>
+     * @OA\Post(
+     *     path="/api/v1/brand/create",
+     *     summary="Create a new brand",
+     *     tags={"Brands"},
+     *     @OA\RequestBody(
+     *         @OA\MediaType(
+     *             mediaType="application/json",
+     *             @OA\Schema(
+     *                 @OA\Property(
+     *                     property="slug",
+     *                     type="string"
+     *                 ),
+     *                 @OA\Property(
+     *                     property="title",
+     *                     oneOf={
+     *                     	   @OA\Schema(type="string"),
+     *                     	   @OA\Schema(type="integer"),
+     *                     }
+     *                 ),
+     *                 example={"slug": "brand-slug", "title": "Brand Title"}
+     *             )
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response="201",
+     *         description="Brand created successfully",
+     *         @OA\JsonContent(
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 description="A message indicating that the brand was created successfully."
+     *             )
+     *         )
+     *     )
+     * )
      */
     public function store(BrandRequest $request)
     {
@@ -49,41 +106,104 @@ class BrandController extends Controller
         // Create an empty brand
         $brand = Brand::make();
 
-        DB::transaction(function () use($data, $brand) {
-           // create a new brand from the validated data
-           $brand->create($data);
+        DB::transaction(function () use ($data, $brand) {
+            // create a new brand from the validated data
+            $brand->create($data);
         });
 
         // Return response with message and data
         return response()->json([
-            'message' => 'Brand created successfully!',
-            'brand' => $brand
+            'message' => 'Brand created successfully!'
         ], Response::HTTP_CREATED); // 201 response for created
     }
 
     /**
-     * Display the specified resource.
-     * 
-     * @author Nderi Kamau <nderikamau1212@gmail.com>
+     * @OA\Get(
+     *     path="/api/v1/brand/{uuid}",
+     *     summary="Get a single brand by UUID",
+     *     tags={"Brands"},
+     *     @OA\Parameter(
+     *         name="uuid",
+     *         in="path",
+     *         description="UUID of the brand to get",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="string",
+     *             format="uuid"
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response="200",
+     *         description="Successful operation",
+     *         @OA\JsonContent(
+     *             @OA\Property(
+     *                 property="brand"
+     *             )
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response="404",
+     *         description="Brand not found"
+     *     )
+     * )
      */
-    public function show(Brand $brand)
+    public function show($uuid)
     {
+        $brand = Brand::where('uuid', $uuid)->first();
+
+        if (!$brand) {
+            return response()->json([
+                'error' => 'Brand not found'
+            ], Response::HTTP_NOT_FOUND);
+        }
+
         return response()->json([
             'brand' => $brand
         ], Response::HTTP_OK);
     }
 
     /**
-     * Update the specified resource in storage.
-     * 
-     * @author Nderi Kamau <nderikamau1212@gmail.com>
+     * @OA\Put(
+     *     path="/api/v1/brand/{uuid}",
+     *     summary="Update a brand",
+     *     tags={"Brands"},
+     *     @OA\Parameter(
+     *         name="uuid",
+     *         in="path",
+     *         description="The UUID of the brand to update",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="string",
+     *             format="uuid"
+     *         )
+     *     ),
+     *     @OA\RequestBody(
+     *         required=true
+     *     ),
+     *     @OA\Response(
+     *         response="200",
+     *         description="Brand updated successfully"
+     *     ),
+     *     @OA\Response(
+     *         response="404",
+     *         description="Brand not found"
+     *     )
+     * )
      */
-    public function update(BrandRequest $request, Brand $brand)
+    public function update(BrandRequest $request, $uuid)
     {
+        $brand = Brand::where('uuid', $uuid)->first();
+
+        if (!$brand) {
+            return response()->json([
+                'error' => 'Brand not found'
+            ], Response::HTTP_NOT_FOUND);
+        }
+
         //validate data
         $data = $request->validated();
 
-        // Update the brand data
+        // Update the $brand data
         $brand->update($data);
 
         //Return response with message and data
@@ -93,17 +213,45 @@ class BrandController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
-     * 
-     * @author Nderi Kamau <nderikamau1212@gmail.com>
+     * @OA\Delete(
+     *     path="/api/v1/brand/{uuid}",
+     *     summary="Delete a brand",
+     *     tags={"Brands"},
+     *     @OA\Parameter(
+     *         name="uuid",
+     *         in="path",
+     *         description="The UUID of the brand to delete",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="string",
+     *             format="uuid"
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response="200",
+     *         description="Brand deleted successfully"
+     *     ),
+     *     @OA\Response(
+     *         response="404",
+     *         description="Brand not found"
+     *     )
+     * )
      */
-    public function destroy(Brand $brand)
+    public function destroy($uuid)
     {
+        $brand = Brand::where('uuid', $uuid)->first();
+
+        if (!$brand) {
+            return response()->json([
+                'error' => 'Brand not found'
+            ], Response::HTTP_NOT_FOUND);
+        }
+
         //  Delete the brand
         $brand->delete();
 
         // Return response with message
-        return redirect()->back()->with([
+        return response()->json([
             'success' => 'Brand deleted successfully!'
         ], Response::HTTP_OK);
     }
